move utilities into connections
